+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+- RCSPF-120742: allowing to copy additional message headers upon wrapping response function
 
 ## 1.0.0 - 2021-03-19
 ### Changed